chore: initial work to add more proper error handling
In this commit I wanted to use `require` to import the `.github/ISSUE_TEMPLATE` in order to include
the file contents.

However, despite doing a lot of research and trying my best to figure out the METRO bundler, I
cannot seem to get this working as-expected. I keep getting a "SHA-1 for file is not computed"
error. According to this thread:

https://github.com/facebook/metro/issues/330

One potential solution is to patch metro in order to generate the SHA. However, I don't want to
patch metro, that admittedly feels a bit flaky as a solution.

In the next commit we'll revert the metro changes and use the GH API to get the bug report file
contents instead