chore: migrate bug reporting to grab template from GH API
Welp, we hit another wall. In this case, we ran into an issue with the GitHub website telling us
that our requested URL is too long. Because of this, this method of opening a template issue
with stacktrace is a no-go.

However, we _can_ use GitHub API to create an issue using `POST` of an arbitrary length
and simply redirect users to the issue `#/edit` and allow them to add details that way

Next commit will update with those values