refactor: remove duplicate color provider
Finally figured out what was happening here.

While it may not be obvious, all of the extended FAB work was being rendered in a
NavigationAwarePortal. Because of this, when we didn't have a Portal.Host, it was raising
the components to a level in the DOM tree without access to the dark mode hooks.
When we added Portal.Host below the dark mode context in 77b81fe512e47ba5fa1ab6d58a26be54cd6917e7
we inadvetertantly (but correctly) also fixed this behavior as well