chore: initial work to remove `isomorphic-git` from our codebase
This is not an indicator of anything negative about the library. Mainly we're looking to remove
`react-native-fs` from our codebase (due to linking issues on Windows) and are currently not
utilizing `isomorphic-git` as much as we could be. Because of this, we're looking to implement
all Git services in native code

For more context: https://oceanbit.dev/blog/gitshark-android-beta/